pyproject.toml
[tool.poetry]
name = "lab-django-superheroes-2023-1113"
version = "0.1.0"
description = ""
authors = ["Vitaliy Kononenko <konvio@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
